Logs/metagpt/part3-2/instruction42/isr_drone_object_detection_tracking/requirements.txt
# Core dependencies
numpy>=1.21.0
opencv-python>=4.5.5.64
PyYAML>=6.0
torch>=2.0.0
ultralytics>=8.0.0

# DeepSORT dependencies
deep_sort_realtime>=1.3.0

# ByteTrack dependencies
bytetrack>=0.4.0

# TensorRT (Jetson Orin NX, install via NVIDIA SDK Manager or pip if available)
tensorrt>=8.6.0

# Optional: PyCUDA for custom GPU ops (if needed)
pycuda>=2022.1

# For logging and system utilities
logging

# For argument parsing
argparse

# For JSON operations
json

# For threading and signal handling (standard library, no pip install needed)

# Note:
# - For TensorRT, it is recommended to use NVIDIA's pre-installed version on Jetson Orin NX.
# - For torch, use the Jetson-compatible wheel from NVIDIA or official PyTorch for ARM64.
# - For ultralytics YOLOv8, ensure compatibility with your torch version.
# - Some dependencies (e.g., logging, argparse, json, threading, signal) are part of Python's standard library and do not require installation via pip.